+++ packages/atomic/src/components/result-template-components/atomic-result-html/atomic-result-html.stories.tsx
@@ -1,9 +1,0 @@
-import defaultResultComponentStory from '../../../../.storybook/default-result-component-story';
-
-const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
-  'Atomic/ResultList/ResultHtml',
-  'atomic-result-html',
-  {field: 'excerpt'}
-);
-export default defaultModuleExport;
-export const DefaultResultHtml = exportedStory;