+++ packages/atomic/src/components/search/atomic-query-summary/atomic-query-summary.stories.tsx
@@ -1,0 +1,9 @@
+import defaultStory from '../../../../.storybook/default-story';
+
+const {defaultModuleExport, exportedStory} = defaultStory(
+  'Atomic/QuerySummary',
+  'atomic-query-summary',
+  {}
+);
+export default defaultModuleExport;
+export const DefaultQuerySummary = exportedStory;