+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.11.2...@coveo/atomic-react@1.11.3) (2022-06-20)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.11.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.11.1...@coveo/atomic-react@1.11.2) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.11.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.11.0...@coveo/atomic-react@1.11.1) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 # [1.11.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.10.0...@coveo/atomic-react@1.11.0) (2022-06-17)